fix: apply PR comment changes
