+++ src/Components/CreateQuiz/FormContainer/FormSteps/KeywordsForm.tsx
@@ -2,7 +2,7 @@ import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 import TextField from "@mui/material/TextField";
 import { useAtom } from "jotai";
-import { newQuizDataAtom } from "../../../Views/CreateQuiz/atoms";
+import { newQuizDataAtom } from "../../../../Views/CreateQuiz/atoms";
 import { Box } from "@mui/material";
 import { useState, useRef, useEffect } from "react";
 import { KeywordTag } from "./KeywordTag";
