+++ src/Components/CreateQuiz/FormContainer/FormSteps/MetaPromptForm.tsx
@@ -3,7 +3,7 @@ import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 import TextField from "@mui/material/TextField";
 import { useAtom } from "jotai";
-import { newQuizDataAtom } from "../../../Views/CreateQuiz/atoms";
+import { newQuizDataAtom } from "../../../../Views/CreateQuiz/atoms";
 import FormStepTitle from "./FormStepTitle";
 
 export default function MetaPromptForm() {
