+++ src/Components/CreateQuiz/FormContainer/FormSteps/QuizInfoForm.tsx
@@ -3,7 +3,7 @@ import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 import TextField from "@mui/material/TextField";
 import { useAtom } from "jotai";
-import { newQuizDataAtom } from "../../../Views/CreateQuiz/atoms";
+import { newQuizDataAtom } from "../../../../Views/CreateQuiz/atoms";
 import FormStepTitle from "./FormStepTitle";
 
 export default function QuizInfoForm() {
@@ -21,7 +21,7 @@ export default function QuizInfoForm() {
       <FormStepTitle>Quiz Info</FormStepTitle>
 
       <Grid container spacing={5}>
-        <Grid item xs={12} sm={12}>
+        <Grid item xs={12}>
           <TextField
             required
             id="quizTitle"
@@ -34,7 +34,7 @@ export default function QuizInfoForm() {
             onChange={handleChange}
           />
         </Grid>
-        <Grid item xs={12} sm={12}>
+        <Grid item xs={12}>
           <TextField
             id="description"
             name="Quiz Description"
