+++ src/Views/CreateQuiz/index.tsx
@@ -1,24 +1,26 @@
-import { useState } from "react";
-import CssBaseline from "@mui/material/CssBaseline";
-import AppBar from "@mui/material/AppBar";
-import Box from "@mui/material/Box";
 import Container from "@mui/material/Container";
-import Toolbar from "@mui/material/Toolbar";
 import Paper from "@mui/material/Paper";
 import Stepper from "@mui/material/Stepper";
 import Step from "@mui/material/Step";
 import StepLabel from "@mui/material/StepLabel";
-import Button from "@mui/material/Button";
-import Link from "@mui/material/Link";
 import Typography from "@mui/material/Typography";
 import DefaultLayout from "../../Layouts/DefaultLayout";
-import CreateQuizForm from "../../Components/CreateQuiz/CreateQuizForm";
+import CreateQuizFormContainer from "../../Components/CreateQuiz/FormContainer";
 import { useAtom } from "jotai";
 import { activeStepAtom, stepNamesAtom, newQuizDataAtom } from "./atoms";
+import FormActions from "../../Components/CreateQuiz/FormContainer/FormActions";
+import { useCreateQuiz } from "../../Api/quizzes";
+import {
+  LoadingBackdrop,
+  SuccessBackdrop,
+} from "../../Components/CreateQuiz/FormContainer/Backdrops/LoadingBackdrop";
 
 export default function CreateQuiz() {
   const [activeStep, setActiveStep] = useAtom(activeStepAtom);
   const [stepNames] = useAtom(stepNamesAtom);
+  const [newQuizData] = useAtom(newQuizDataAtom);
+
+  const { mutate, isPending, isSuccess } = useCreateQuiz();
 
   return (
     <DefaultLayout>
@@ -30,10 +32,15 @@ export default function CreateQuiz() {
           variant="outlined"
           sx={{
             my: { xs: 3, md: 4 },
-            p: { xs: 2, md: 3 },
+            p: { xs: 2, md: 5 },
             minWidth: "30rem",
           }}
         >
+          {/* Quiz creation request in progress */}
+          {isPending && <LoadingBackdrop isPending={isPending} />}
+          {/* Quiz creation request successful */}
+          {isSuccess && <SuccessBackdrop isSuccess={isSuccess} />}
+
           <Stepper activeStep={activeStep} sx={{ pt: 3, pb: 5 }}>
             {stepNames.map((label) => (
               <Step key={label}>
@@ -41,7 +48,8 @@ export default function CreateQuiz() {
               </Step>
             ))}
           </Stepper>
-          <CreateQuizForm />
+          <CreateQuizFormContainer />
+          <FormActions createQuiz={() => mutate(newQuizData)} />
         </Paper>
       </Container>
     </DefaultLayout>
